+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Support for predict Algos (#1071)
+- Hardcode task outputs (#1068)
 
 ### Changed